+++ package.json
@@ -1,0 +1,7 @@
+{
+  "name": "ExpertMark",
+  "version": "1.0.0",
+  "dependencies": {
+    "jquery": "^3.6.0"
+  }
+}
