+++ views/createDelphiQuiz.html
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>CreateDelphiQuizForecast</title>
+    <script src="script/createDelphiQuizScript.js"></script>
+</head>
+<body>
+<header class="header">
+    <div class="container">
+        <div class="header__inner">
+            <div class="header_logo">
+                <h1>ExpertMark</h1>
+            </div>
+        </div>
+    </div>
+</header>
+
+<!--Box-->
+<div id="forecast_data" class="box">
+    <div class="header__inner">
+        Create Delphi Quiz
+    </div>
+    <div class="header__inner">
+        <div class="header__inner">Title</div>
+        <label>
+            <input id="title" type="text">
+        </label>
+        <div class="header__inner">Asset Name</div>
+        <label>
+            <input id="asset_name" type="text">
+        </label>
+        <div class="header__inner">Description</div>
+        <label>
+            <input id="description" type="text">
+        </label>
+        <div class="header__inner">DiscussionTimeInDays</div>
+        <label>
+            <input id="discussion_time_in_days" type="number">
+        </label>
+        <div class="header__inner">ExpertUsernames</div>
+        <label>
+            <input id="expert_usernames" type="text">
+        </label>
+    </div>
+    <div id="send_delphi_quiz_button" class="btn">
+        <button>
+            Create Forecast
+        </button>
+    </div>
+</div>
+
+<!--Menu-->
+<div class="menu">
+    <div class="container">
+        <div class="menu__inner">
+            <a class="btn  btn--red" href="#">Profile</a>
+            <a class="btn  btn--red" href="#">My Forecasts</a>
+            <a class="btn  btn--red" href="#">Assets</a>
+            <a class="btn  btn--red" href="#">Delphi</a>
+        </div>
+    </div>
+</div>
+
+<!--Footer-->
+<footer class="footer">
+    <div class="copyright">
+        <div class="container">
+            <div class="copyright__text">
+                <div>Copyright © 2022 ExpertMark. All Rights Reserved</div>
+                <div>Made with <span style="color:#e84545;">&#10084</span> <span>by Mihail Nenahov</span></div>
+            </div>
+        </div>
+    </div>
+</footer>
+
+</body>
+</html>
