+++ views/index.html
@@ -23,7 +23,7 @@
                     </div>
 
                     <nav class="nav">
-                        <a href="signin.html" class="nav__link">Sign In</a>
+                        <a href="signIn.html" class="nav__link">Sign In</a>
                         <a href="#" class="nav__link">Sign Out</a>
                     </nav>
                 </div>
