+++ views/script/signIn.js
@@ -1,6 +1,4 @@
-let script = document.createElement('script');
-script.src = 'https://code.jquery.com/jquery-3.6.0.min.js';
-document.getElementsByTagName('head')[0].appendChild(script);
+import $ from "jquery"
 
 $("#sign_in_button").click(function() {
     let username = $("username").val();
